Add a forgotten static_cast

+++ breath/environment/brt/dep/syst/windows/node_id.cpp
@@ -75,7 +75,7 @@ operator <<( std::ostream & os, node_id const & id )
     os.setf( std::ios_base::hex, std::ios_base::basefield ) ;
 
     for ( auto it = id.m_address.cbegin(); it != id.m_address.cend() ; ++ it ) {
-        os << std::setw( 2 ) << *it ;
+        os << std::setw( 2 ) << static_cast< int >( *it ) ;
         if ( it != ( id.m_address.cend() - 1 )) {
             os << '-' ;
         }
